requirements.txt
# Airflow
apache-airflow==2.8.1
apache-airflow-providers-ssh==3.10.0

# Spark
pyspark==3.5.0

# AWS
boto3==1.34.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Linting
pylint==3.0.3
black==23.12.1